Fixed the cursor
