+++ libsrc/_DEVELOPMENT/EXAMPLES/vgl/emu.bat
@@ -3,10 +3,12 @@ SET BINPATH=%CD%
 SET MESSPATH=Z:\Apps\_emu\MESSUI-0.181
 SET ROMPATH=Z:\Apps\_emu\_roms
 
-SET PROGNAME=hello
+REM SET PROGNAME=hello
+SET PROGNAME=%1
+
 REM SET EMUSYS=gl2000
-REM SET EMUSYS=gl4000
-SET EMUSYS=gl4004
+SET EMUSYS=gl4000
+REM SET EMUSYS=gl4004
 REM SET EMUSYS=gl5000
 REM SET EMUSYS=gl3000s
