+++ libsrc/_DEVELOPMENT/EXAMPLES/vgl/hello2.c
